requirements.txt
Flask
requests
azure-storage-blob
opencv-python
numpy
imageio-ffmpeg
twilio
python-dotenv
Flask-CORS